Get build working
