+++ service/Makefile.am
@@ -135,9 +135,12 @@ libgeopmd_la_SOURCES = contrib/json11/json11.cpp \
                        src/geopm.h \
                        src/geopm_debug.hpp \
                        src/geopm_error.h \
+                       src/geopm_internal.h \
                        src/geopm_pio.h \
                        src/geopm_plugin.cpp \
                        src/geopm_plugin.hpp \
+                       src/geopm_sched.c \
+                       src/geopm_sched.h \
                        src/geopm_time.h \
                        src/geopm_topo.h \
                        src/geopm_version.c \
@@ -150,14 +153,14 @@ libgeopmd_la_SOURCES = contrib/json11/json11.cpp \
                        # end
 
 if ENABLE_NVML
-    libgeopmpolicy_la_SOURCES += src/NVMLAcceleratorTopo.cpp \
-                                 src/NVMLAcceleratorTopo.hpp \
-                                 src/NVMLDevicePool.cpp \
-                                 src/NVMLDevicePool.hpp \
-                                 src/NVMLDevicePoolImp.hpp \
-                                 src/NVMLIOGroup.cpp \
-                                 src/NVMLIOGroup.hpp \
-                                 #end
+    libgeopmd_la_SOURCES += src/NVMLAcceleratorTopo.cpp \
+                            src/NVMLAcceleratorTopo.hpp \
+                            src/NVMLDevicePool.cpp \
+                            src/NVMLDevicePool.hpp \
+                            src/NVMLDevicePoolImp.hpp \
+                            src/NVMLIOGroup.cpp \
+                            src/NVMLIOGroup.hpp \
+                            #end
 endif
 
 # RPM TARGET
