+++ service/geopm-service.spec.in
@@ -1,0 +1,139 @@
+#  Copyright (c) 2015 - 2021, Intel Corporation
+#
+#  Redistribution and use in source and binary forms, with or without
+#  modification, are permitted provided that the following conditions
+#  are met:
+#
+#      * Redistributions of source code must retain the above copyright
+#        notice, this list of conditions and the following disclaimer.
+#
+#      * Redistributions in binary form must reproduce the above copyright
+#        notice, this list of conditions and the following disclaimer in
+#        the documentation and/or other materials provided with the
+#        distribution.
+#
+#      * Neither the name of Intel Corporation nor the names of its
+#        contributors may be used to endorse or promote products derived
+#        from this software without specific prior written permission.
+#
+#  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+#  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+#  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+#  A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+#  OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+#  SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+#  LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+#  DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+#  THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+#  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY LOG OF THE USE
+#  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+#
+Summary: Global Extensible Open Power Manager Service
+Name: geopm-service
+Version: @VERSION@
+Release: 1
+License: BSD-3-Clause
+Group: System Environment/Libraries
+Vendor: Intel Corporation
+URL: https://geopm.github.io
+Source0: geopm-service.tar.gz
+BuildRoot: %{_tmppath}/geopm-%{version}-%{release}-root
+BuildRequires: gcc-c++
+BuildRequires: unzip
+BuildRequires: libtool
+
+%if 0%{?rhel} >= 8
+# Needed to generate debuginfo packages
+BuildRequires: gdb-headless
+%endif
+
+%if ((%{defined suse_version}) && (0%{?suse_version} < 1500)) || ((%{defined rhel}) && (0%{?rhel} < 8))
+# The version of python 3 available in these versions is not supported for all
+# of our package dependencies, but they do support 2.7 and newer versions of 3
+# We originally released to these distributions with an implicit python 2
+# dependency, so keep it that way for existing compatibility.
+%define python_major_version %{nil}
+BuildRequires: python
+BuildRequires: python-devel
+%define python_bin /usr/bin/python
+%else
+%define python_major_version 3
+%if %{defined rhel}
+# We should depend on python36, but that package is implemented as a module.
+# The enterprise Linux repos on opensuse's OBS instance don't expose modules
+# at this time, so the build fails due to a seemingly invalid dependency.
+# For now, depend on platform-python instead, which is a bare package.
+BuildRequires: platform-python
+BuildRequires: platform-python-devel
+%else
+BuildRequires: python3
+BuildRequires: python3-devel
+%endif
+%define python_bin %{__python3}
+%endif
+
+%if 0%{?suse_version} >= 1320
+BuildRequires: openssh
+%endif
+
+Prefix: %{_prefix}
+
+%if %{defined suse_version}
+%define docdir %{_defaultdocdir}/geopm
+%else
+%define docdir %{_defaultdocdir}/geopm-%{version}
+%endif
+
+%description
+The service package for GEOPM.
+
+%prep
+
+%setup -n geopm-%{version}
+
+%package -n python%{python_major_version}-geopmdpy
+Summary: Global Extensible Open Power Manager Service - python %{python_major_version}
+Group: System Environment/Libraries
+Requires: geopm-service
+%{?python_provide:%python_provide python%{python_major_version}-geopmdpy}
+
+%description -n python%{python_major_version}-geopmdpy
+Python %{python_major_version} package for GEOPM.
+
+%build
+test -f configure || ./autogen.sh
+
+./configure --prefix=%{_prefix} --libdir=%{_libdir} --libexecdir=%{_libexecdir} \
+            --includedir=%{_includedir} --sbindir=%{_sbindir} \
+            --mandir=%{_mandir} --docdir=%{docdir} \
+            --with-python=%{python_bin} \
+            || ( cat config.log && false )
+
+%{__make} %{?_smp_mflags}
+
+%install
+%{__make} DESTDIR=%{buildroot} install
+rm -f $(find %{buildroot}/%{_libdir} -name '*.a'; \
+        find %{buildroot}/%{_libdir} -name '*.la')
+
+%clean
+
+%post
+/sbin/ldconfig
+
+%preun
+
+%postun
+/sbin/ldconfig
+
+%files
+%defattr(-,root,root,-)
+%{_libdir}/libgeopmd.so.0.0.0
+%{_libdir}/libgeopmd.so.0
+%{_libdir}/libgeopmd.so
+%dir %{_libdir}/geopm
+%{_bindir}/geopmaccess
+%dir %{docdir}
+
+%files -n python%{python_major_version}-geopmpy
+%{expand:%{python%{python_major_version}_sitelib}}/*
