+++ service/src/IOGroup.cpp
@@ -39,8 +39,6 @@
 #include "MSRIOGroup.hpp"
 #include "CpuinfoIOGroup.hpp"
 #include "TimeIOGroup.hpp"
-#include "ProfileIOGroup.hpp"
-#include "EpochIOGroup.hpp"
 #include "SSTIOGroup.hpp"
 #include "Helper.hpp"
 #include "config.h"
@@ -100,10 +98,6 @@ namespace geopm
                         TimeIOGroup::make_plugin);
         register_plugin(CpuinfoIOGroup::plugin_name(),
                         CpuinfoIOGroup::make_plugin);
-        register_plugin(ProfileIOGroup::plugin_name(),
-                        ProfileIOGroup::make_plugin);
-        register_plugin(EpochIOGroup::plugin_name(),
-                        EpochIOGroup::make_plugin);
         register_plugin(SSTIOGroup::plugin_name(),
                         SSTIOGroup::make_plugin);
 #ifdef GEOPM_CNL_IOGROUP
