+++ service/src/MSRIOGroup.cpp
@@ -46,8 +46,7 @@
 #include "contrib/json11/json11.hpp"
 
 #include "geopm_sched.h"
-#include "geopm_hash.h"
-#include "Environment.hpp"
+#include "geopm_internal.h"
 #include "Exception.hpp"
 #include "Agg.hpp"
 #include "MSRIOImp.hpp"
@@ -795,12 +794,7 @@ namespace geopm
     {
         std::set<std::string> data_files;
         // search path for additional json files to parse
-        std::string env_plugin_path = environment().plugin_path();
         std::vector<std::string> plugin_paths {GEOPM_DEFAULT_PLUGIN_PATH};
-        if (!env_plugin_path.empty()) {
-            std::vector<std::string> dirs = string_split(env_plugin_path, ":");
-            plugin_paths.insert(plugin_paths.end(), dirs.begin(), dirs.end());
-        }
         std::vector<std::unique_ptr<MSR> > msr_arr_custom;
         for (const auto &dir : plugin_paths) {
             auto files = list_directory_files(dir);
